Use tagged version of php-sdk-binary-tools (#5)
See <https://github.com/php/php-sdk-binary-tools?tab=readme-ov-file#internal-notes>.

According to that documentation, for PHP 7.0 and 7.1, PHP SDK 2.1 would
be required; however, we already used newer versions, and so far nobody
had complained, apparently.  So maybe the php-sdk-binary-tools README
is not quite correct, or we would need to adapt the tag for those PHP
versions.  On the other hand, it is somewhat unlikely that these old
versions are still in use, so we don't care too much about that.