Close #8: Implement caching (#14)
For now we cache only the PHP SDK, the PHP binaries and the development
packs.  This already greatly improves the setup performance (it might
easily safe a minute or two, in case of cache hits).  To be able to use
the current PHP revision as part of the cache key, we factor out
determine-revision.ps1.  We create a separate cache for the PHP-SDK
since this likely rarely changes, and since the cached variant is
apparently much faster than fetching a GH release of the PHP-SDK.

Since clients may not want to use the cache, possibly because they have
already a lot of other files in their caches, we explicitly require
clients to opt-in via the `cache` input parameter.